Refactors sass and component folder structure
