+++ README.md
@@ -1,9 +1,30 @@
 # blogwise-landing
 
-## Building
+This site is built using [Gatsby.js](https://www.gatsbyjs.org/). It includes the
+following pages:
 
-`npm install`
+- Home at `/`
+- Features at `/features`
+- Pricing at `/pricing`
+- Privacy Policy at `/privacy-policy`
 
-## Running
+## Getting Started
 
- `npm start`
+### Setup
+
+After you have setup your work environment as outlined in the
+[blogwise-lib](https://github.com/nputnam-hu/blogwise-lib), all you need to do
+is clone this repo and run `npm install`.
+
+### Development
+
+Be sure to read through the [Contribution section in
+blogwise-lib](https://github.com/nputnam-hu/blogwise-lib/blob/master/README.md#contribution).
+
+To start the dev server, just run `npm start`.
+
+### Deployment
+
+This website is deployed using Netlify's CI. This means that a preview is built
+as a part of every open PR, and a production version is deployed when the PR is
+merged into master.
