+++ src/components/Footer/Footer.module.sass
@@ -1,0 +1,44 @@
+.footer
+  background: #F4F4F4
+  padding-bottom: 20px
+  position: absolute
+  width: 100vw
+  margin-top: 40px
+  bottom: 0
+  padding: 30px 40px
+  display: flex
+  flex-direction: column
+  overflow-x: hidden
+  @media (max-width: 425px)
+    width: auto
+
+.footer__bottom
+  display: flex
+  flex-direction: row
+  justify-content: space-between
+  @media (max-width: 800px)
+    flex-direction: column-reverse
+
+.footer__links
+  display: flex
+  flex-direction: row
+  align-self: flex-start
+  margin-top: 30px
+  font-family: Nunito
+  font-size: 13px
+  color: #6D6D6D
+  span
+    margin-right: 60px
+  a
+    text-decoration: none
+    color: #6D6D6D
+    margin-right: 30px
+
+.footer__social
+  margin-top: -30px
+  display: flex
+  flex-direction: row
+  align-self: flex-end
+  margin-right: 70px
+  a
+    padding-right: 15px
