+++ src/components/Layout/Layout.module.sass
@@ -1,0 +1,5 @@
+@import ../../styles/global
+
+.layout
+  overflowX: hidden
+  overflowY: scroll !important
