+++ src/components/Navbar/Navbar.module.sass
@@ -1,0 +1,58 @@
+.navbar
+  padding: 20px 25px
+  display: flex
+  flex-direction: row
+  align-items: center
+  justify-content: space-between
+  @media (max-width: 900px)
+    background: #67b4f9
+
+  nav
+    padding-right: 20px
+    @media (max-width: 900px)
+      display: none
+
+.navbar__link
+  margin-right: 50px
+  text-transform: lowercase
+  @media (max-width: 900px)
+      margin-right: 50px
+      text-transform: lowercase
+      font-size: 18px
+      font-weight: 600
+      text-decoration: none
+      color: black
+      font-family: Nunito
+
+.buttonBlue
+  text-decoration: none
+  margin-left: 10px
+  padding: 13px 35px
+  font-weight: 300
+  font-size: 16px
+  border-radius: 30px
+  background-color: #75B1E1
+  color: white
+  text-transform: uppercase
+  letter-spacing: 1.6px
+
+.bm-burger-bars
+  background: white
+
+.mobile
+  display: none !important
+  @media (max-width: 900px)
+    display: inline !important
+    font-size: 22px
+
+.bm-menu-wrap
+  overflow-y: hidden
+
+.navbar-link-mobile
+  text-transform: lowercase
+  text-decoration: none
+  font-family: Nunito
+  font-size: 24px
+  font-weight: 600
+  margin-top: 20px
+  color: #333
