+++ src/components/WhyCard/WhyCard.module.sass
@@ -1,4 +1,4 @@
-.whycard-container
+.whyCard
   width: 475px
   padding: 30px 70px 50px
   height: 250px
@@ -14,22 +14,22 @@
     width: 80%
     padding: 10px 10% 10px
     height: 520px
-  .whycard-icon
-    margin-bottom: 20px
-    height: 20px
-  .whycard-text
-    width: 275px
-    display: flex
-    flex-direction: column
-    font-family: Nunito
-    @media (max-width: 600px)
-      width: 95%
-    h2
-      text-align: center
-      text-transform: lowercase
-      font-size: 34px
-      margin-bottom: -1px
-      padding-bottom: 15px
+.whyCard__icon
+  margin-bottom: 20px
+  height: 20px
+.whyCard__text
+  width: 275px
+  display: flex
+  flex-direction: column
+  font-family: Nunito
+  @media (max-width: 600px)
+    width: 95%
+  h2
+    text-align: center
+    text-transform: lowercase
+    font-size: 34px
+    margin-bottom: -1px
+    padding-bottom: 15px
 
-.whycard-container:hover
+.whyCard:hover
   box-shadow: 6px 6px 8px 2px #888888
