+++ src/components/WhyCard/index.js
@@ -1,15 +1,3 @@
-import React from 'react'
-import Img from 'gatsby-image'
-import './styles.sass'
-
-const WhyCard = ({ title, description, iconFixed }) => (
-  <div className="whycard-container">
-    <Img fixed={iconFixed} alt={title} className="whycard-icon" />
-    <div className="whycard-text">
-      <h2>{title}</h2>
-      <p>{description}</p>
-    </div>
-  </div>
-)
+import WhyCard from './WhyCard'
 
 export default WhyCard
