+++ src/components/WhyCardGrid/WhyCardGrid.module.sass
@@ -1,4 +1,4 @@
-.whycard-row
+.whyCardRow
   display: flex
   flex-direction: row
   justify-content: space-between
