+++ src/pages/pricing.js
@@ -1,8 +1,10 @@
 import React from 'react'
 import { graphql } from 'gatsby'
 import Img from 'gatsby-image'
+
 import Layout from '../components/Layout'
-import lightning from './lightning.png'
+
+import lightning from '../images/small-lightning.png'
 import styles from '../styles/pricing.module.sass'
 
 const PricingCard = ({
@@ -14,16 +16,16 @@ const PricingCard = ({
   branding,
   description,
 }) => (
-  <div className={styles.pricingcard}>
-    <div className={styles.pricingcard__top} style={{ background: bgColor }}>
+  <div className={styles.pricingCard}>
+    <div className={styles.pricingCard__top} style={{ background: bgColor }}>
       <h2>{name}</h2>
     </div>
-    <div className={styles.pricingcard__bottom}>
+    <div className={styles.pricingCard__bottom}>
       <h1>${price}</h1>
-      <span className={styles.pricingcard__permonth}>
+      <span className={styles.pricingCard__permonth}>
         per month, billed monthly
       </span>
-      <ul className={styles.pricingcard__list}>
+      <ul className={styles.pricingCard__list}>
         <li>
           <img src={lightning} alt="Bullet Point" />
           <span>
@@ -40,7 +42,7 @@ const PricingCard = ({
           <span>{branding ? 'blogwise branding' : 'ad free'}</span>
         </li>
       </ul>
-      <span className={styles.pricingcard__description}>{description}</span>
+      <span className={styles.pricingCard__description}>{description}</span>
     </div>
   </div>
 )
@@ -61,7 +63,7 @@ const Pricing = ({ data }) => (
     />
     <div className={styles.pricing}>
       <h1>Pricing Options</h1>
-      <div className={styles.pricingcards}>
+      <div className={styles.pricingCards}>
         <PricingCard
           bgColor="rgba(204, 204, 204, 0.42)"
           name="Free"
@@ -96,10 +98,10 @@ const Pricing = ({ data }) => (
           description="for businesses killing the marketing game"
         />
       </div>
-      <div className={styles.sectiontwo}>
+      <div className={styles.getStartedSection}>
         <h1>Not sure what plan is best for you?</h1>
         <span>Start with a free blog and upgrade later</span>
-        <a href="https://app.blogwise.co" className="yellow-button">
+        <a href="https://app.blogwise.co" className={styles.yellowButton}>
           Get Started
         </a>
       </div>
