+++ src/styles/features.module.sass
@@ -1,0 +1,203 @@
+@import global
+
+.features
+  padding-bottom: 100px
+
+////////////////////////
+// Individual HowCard //
+////////////////////////
+.howCard
+  display: flex
+  flex-direction: column
+  align-items: center
+  padding-right: 50px
+  @media (max-width: 1120px)
+    margin-bottom: 50px
+    flex-direction: row
+.howCard__top
+  display: flex
+  flex-direction: row
+  @media (max-width: 1120px)
+    flex-direction: column
+  @media (max-width: 600px)
+    margin-left: -100px
+.howCard__stepcounter
+  margin-left: 24px !important
+  color: #d6d6d6
+  margin-bottom: 2px
+  @media (max-width: 1120px)
+    font-size: 40px
+    margin-right: 20px
+    margin-left: 0px
+    text-align: right
+.howCard__dots
+  position: absolute
+  margin-left: 50px
+  width: 375px
+  @media (max-width: 1120px)
+    position: relative
+    margin-left: 0px
+    margin-top: 150px
+    margin-right: -145px
+    height: 25px
+    transform: rotate(270deg)
+.howCard__bottom
+  width: 350px
+  height: 400px
+  box-shadow: 3px 3px 4px 2px #E9E9E9
+  display: flex
+  flex-direction: column
+  align-items: center
+  @media (max-width: 600px)
+    width: 275px
+    height: 350px
+  span
+    margin-top: 30px
+    font-size: 18px
+    font-weight: 600
+  p
+    margin-top: auto
+    width: 90%
+    font-size: 15px
+    @media (max-width: 600px)
+      font-size: 14px
+.howCard__img
+  margin-top: auto
+  margin-bottom: auto
+  width: 200px
+  @media (max-width: 600px)
+    width: 175px
+
+////////////////////////
+// Individual Feature //
+////////////////////////
+.feature
+  display: flex
+  width: 100%
+  justify-content: space-around
+  @media (max-width: 1200px)
+    flex-direction: column !important
+    align-items: center
+    justify-content: flex-start
+.feature__img  
+  width: 500px
+  height: 470px
+  background: white
+  margin: 20px
+  box-shadow: 3px 3px 4px 2px #E9E9E9
+  @media (max-width: 600px)
+    width: 350px
+    height: 315px
+.feature__text
+  width: 400px
+  display: flex
+  flex-direction: column
+  margin-top: 20px
+  @media (max-width: 1200px)
+    width: 60%
+  @media (max-width: 600px)
+    width: 95%
+  h1
+    margin-bottom: 40px
+    @media (max-width: 600px)
+      margin-bottom: 0px
+  p
+    font-size: 18px
+
+////////////////////
+// Header Section //
+////////////////////
+.header
+  padding: 40px 12% 200px
+  height: 400px
+  width: 500px
+  display: flex
+  flex-direction: column
+  justify-content: center
+  @media (max-width: 1200px)
+    padding: 40px 5% 200px
+  @media (max-width: 600px)
+    padding-bottom: 100px
+    width: 90%
+  h1
+    font-family: Nunito
+    font-size: 40px
+    padding-bottom: 10px
+  p
+    width: 400px
+    font-size: 18px
+    margin-top: -1px
+    @media (max-width: 600px)
+      width: 90%
+.header__photo
+  top: 150px
+  right: 10%
+  height: 450px
+  width: 470px
+  @media (max-width: 1100px)
+    right: 20px
+  @media (max-width: 1000px)
+    display: none !important
+.yellowButton
+  @extend .yellowButton
+  margin-top: 30px
+  width: 157px
+
+//////////////////////////
+// How It Works Section //
+//////////////////////////
+.howItWorksSection
+  padding: 0px 12% 100px
+  display: flex
+  flex-direction: column
+  align-items: center
+.howItWorksSection__header
+  font-size: 28px
+  margin-bottom: 50px
+.howItWorksSection__howCards
+  display: flex
+  flex-direction: row
+  @media (max-width: 1120px)
+    flex-direction: column
+
+//////////////////////
+// Features Section //
+//////////////////////
+.featuresSection
+  padding: 40px 10% 100px
+  background: #F4F4F4
+  display: flex
+  flex-direction: column
+  align-items: center
+  @media (max-width: 1200px)
+    padding: 40px 4% 100px
+.featuresSection__title
+  text-transform: uppercase
+  font-size: 26px
+  font-weight: 300
+  margin-bottom: 100px
+
+///////////////////////
+// Companies Section //
+///////////////////////
+.companiesSection
+  padding: 70px 10% 100px
+  display: flex
+  flex-direction: column
+  align-items: center
+  text-align: center
+  h1
+    font-weight: 300
+    margin-bottom: 80px
+.companiesSection__imgs
+  display: flex
+  flex-direction: row
+  justify-content: space-around
+  align-items: center
+  flex-wrap: wrap
+  width: 80%
+  @media (max-width: 800px)
+    flex-direction: column
+    justify-content: flex-start
+.companiesSection__img
+  margin-bottom: 40px
