+++ src/styles/global.sass
@@ -17,3 +17,12 @@ body
   background: #FFD833
   color: black
   text-transform: uppercase
+
+.yellowButton
+  text-decoration: none
+  padding: 12px 45px
+  border-radius: 40px
+  font-weight: 600
+  background: #FFD833
+  color: black
+  text-transform: uppercase
